Content done

+++ app/src/main/java/com/bruno/aybar/composechallenges/backup/BackupUi.kt
@@ -4,7 +4,7 @@ sealed class BackupUi {
 
     data class RequestBackup(val lastBackup: String): BackupUi()
 
-    data class BackupInProgress(val progress: Float): BackupUi()
+    data class BackupInProgress(val progress: Int): BackupUi()
 
     object BackupCompleted: BackupUi()
 }
